# docs/api.yaml
openapi: 3.0.2
info:
  title: M & O bank
  version: 1.0.0
  description: Bank api
  contact:
    email: maarja.tapsi@khk.ee
servers:
  -
    url: 'http://localhost:3000'
paths:
...
  /transactions:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
            examples:
              createTransaction:
                value:
                  accountFrom: EE2860832851762801
                  accountTo: EE2798567388301682
                  amount: 100
                  explanation: test
        required: true
      tags:
        - transactions
      responses:
        '201':
          content:
            application/json:
              examples:
                successfulTransaction:
                  value:
                    message: Transaction created successfully
          description: Transaction created successfully
        '400':
          content:
            application/json:
              examples:
                invalidAmount:
                  value:
                    error: Invalid amount
          description: Invalid amount
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '402':
          content:
            application/json:
              examples:
                insufficientFunds:
                  value:
                    error: Insufficient funds
          description: Insufficent funds
        '403':
          content:
            application/json:
              examples:
                forbidden:
                  value:
                    error: Forbidden
          description: The accountFrom does not belong to the user
        '404':
          content:
            application/json:
              examples:
                accountNotFound:
                  value:
                    error: Account not found
          description: Account not found
      security:
        -
          Authorize: []
  /transactions/b2b:
    post:
      tags:
        - transactions
      responses:
        '200':
          description: Succesfully received transaction
        '400':
          content:
            application/json:
              examples:
                parsingJWTFailed:
                  value:
                    error: 'Parsing JWT payload failed: reason here'
          description: Returns the errors specified
        '404':
          content:
            application/json:
              examples:
                accountNotFound:
                  value:
                    error: Account not found
          description: Account not found
        '500':
          content:
            application/json:
              examples:
                cannotVerifySignature:
                  value:
                    error: 'Cannot verify your signature: The jwksUrl of your bank is missing'
          description: "Cannot verify your signature: The jwksUrl of your bank is missing\r\n"
        '502':
          content:
            application/json:
              examples:
                centralBankError:
                  value:
                    error: 'Central Bank error: reason here'
          description: Central Bank error
...
components:
  schemas:
...
    TransactionRequest:
      title: Root Type for TransactionRequest
      description: ''
      type: object
      properties:
        accountFrom:
          type: string
        accountTo:
          type: string
        amount:
          format: int32
          type: integer
        explanation:
          type: string
      example:
        accountFrom: EE2860832851762801
        accountTo: EE2798567388301682
        amount: 100
        explanation: test
  responses:
    UnauthorizedError:
      content:
        application/json:
          examples:
            Unauthorized:
              value:
                error: Error description here
      description: Access token is missing or invalid
  securitySchemes:
    Authorize:
      scheme: bearer
      type: http